requirements.txt
langchain==0.1.0
python-dotenv==1.0.0
discord==2.3.2